+++ src/index.d.ts
@@ -6,8 +6,13 @@ import {
   Store,
 } from 'redux';
 
+type CustomContext = React.Context<unknown>;
+
+export type createCustomContext = () => CustomContext;
+
 export type ReduxProviderProps<S, A extends Action> = {
   store: Store<S, A>;
+  customContext?: CustomContext;
 };
 
 export type ReduxProviderType<S = unknown, A extends Action = AnyAction>
@@ -17,11 +22,16 @@ export const ReduxProvider: ReduxProviderType;
 
 export const useReduxDispatch: <A extends Action>() => Dispatch<A>;
 
-export const useReduxState: <S extends {}>() => S;
+type Opts = {
+  customContext?: CustomContext;
+};
+
+export const useReduxState: <S extends {}>(opts: Opts) => S;
 
 export const useReduxSelectors: <S extends {}, M extends {}>(
   selectors: { [K in keyof M]: (state: S) => M[K] },
+  opts: Opts,
 ) => M;
 
-export const useReduxStateSimple: <S extends {}>() => S;
-export const useReduxStateRich: <S extends {}>() => S;
+export const useReduxStateSimple: <S extends {}>(opts: Opts) => S;
+export const useReduxStateRich: <S extends {}>(opts: Opts) => S;