+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Direct state context
   - No breaking change in API, but it may behave differently
+- Support custom context
 
 ## [3.0.0] - 2019-05-18
 ### Changed